+++ src/test/java/com/vaadin/ui/grid/tests/GridTestPageIT.java
@@ -13,7 +13,7 @@
  * License for the specific language governing permissions and limitations under
  * the License.
  */
-package com.vaadin.ui.grid;
+package com.vaadin.ui.grid.tests;
 
 import org.junit.Assert;
 import org.junit.Before;
